Experimental work on large datasets
Pulled some code from Network to Train in an attempt
to be able to train large dataset. The main idea was
to stream one example at a time when we compute the
big deltas instead of passing a large matrix of x and y.

We don't like this code enough to merge it yet.